Clean up merge
