+++ Volume/Models/Publication.swift
@@ -16,6 +16,7 @@ struct Publication: Hashable, Identifiable {
     let recent: String
     let shoutouts: Int
     let websiteURL: URL?
+    let socials: [String : String] = [:]
     
     init(
         bio: String,
@@ -26,8 +27,8 @@ struct Publication: Hashable, Identifiable {
         shoutouts: Int,
         websiteURL: URL?
     ) {
-        self.articles = articles
-        self.description = description
+//        self.articles = articles
+        self.bio = bio
         self.name = name
         self.id = id
         self.profileImageURL = imageURL
