+++ Volume/View Models/ArticleRow.swift
@@ -12,6 +12,11 @@ import SwiftUI
 struct ArticleRow: View {
     let article: Article
     let showsPublicationName: Bool
+    
+    init(article: Article, showsPublicationName: Bool = true) {
+        self.article = article
+        self.showsPublicationName = showsPublicationName
+    }
         
     var body: some View {
         HStack(spacing: 20) {
